backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-environ==0.12.0
psycopg2-binary==2.9.10
django-filter==24.2
drf-spectacular==0.28.0
pytest==8.4.2
pytest-django==4.11.1
factory_boy==3.3.3
gunicorn==23.0.0
